==> 00.clean_data.R <==
#--------------------------------------------------------------------------#
# Project: PlanktonDist
# Script purpose: Clean data
# Date: 26/01/2024
# Author: Thelma Panaïotis
#--------------------------------------------------------------------------#

source("utils.R")


## Load raw data ----
#--------------------------------------------------------------------------#
plankton <- read_parquet("data/raw/plankton_predictions.parquet")
cops <- read_parquet("data/raw/copepods.parquet")
images <- read_parquet("data/raw/predicted_images.parquet")


## Refine copepods taxonomy ----
#--------------------------------------------------------------------------#
plankton <- plankton %>% 
  # keep only relevant columns
  select(
    transect, img_name, object_id = id, taxon,
    datetime, lat, lon, dist, depth,
    esd = equivalent_diameter, # ESD
    x = centroid_1, y = centroid_0, # positional data
    contains("bbox") # for height and width
  ) %>% 
  # join with cops
  left_join(cops %>% select(object_id, new_taxon = taxon), by = join_by(object_id)) %>% 
  # keep only new taxonomy if present
  mutate(taxon = ifelse(!is.na(new_taxon), new_taxon, taxon)) %>% 
  select(-new_taxon)

# Read clean taxa names
clean_taxa <- read_csv("data/raw/taxa_list.csv", show_col_types = FALSE) %>% select(taxon = orig_name, new_taxon = new_name)
taxa <- clean_taxa %>% drop_na(new_taxon) %>% pull(new_taxon) %>% sort() %>% unique()

# Add clean taxa names to plankton table
plankton <- plankton %>% 
  left_join(clean_taxa, by = join_by(taxon)) %>% 
  select(-taxon) %>% 
  rename(taxon = new_taxon) %>% 
  # drop objects that should be ignored
  drop_na(taxon)


## Compute height and width ----
#--------------------------------------------------------------------------#
plankton <- plankton %>% 
  mutate(
    width = bbox_3 - bbox_1,
    height = bbox_2 - bbox_0,
    .after = y
  ) %>% 
  select(-contains("bbox"))


## Keep only images with more than 1 object ----
#--------------------------------------------------------------------------#
# names of images with more than 1 objects
img_names <- count(plankton, img_name) %>% filter(n > 1) %>% pull(img_name)

# keep only objects in these images
plankton <- plankton %>% filter(img_name %in% img_names)
images <- images %>% filter(img_name %in% img_names)


## Save clean data ----
#--------------------------------------------------------------------------#
write_parquet(plankton, sink = "data/00.plankton_clean.parquet")
write_parquet(images, sink = "data/00.images_clean.parquet")


## Save a subsample ----
#--------------------------------------------------------------------------#
# Generate a subsample of 100 images for tests
n_img <- 10000
# subsample images
images_sub <- images %>% slice_sample(n = n_img) 
# keep plankton in sampled images
plankton_sub <- plankton %>% filter(img_name %in% images_sub$img_name)

# Save
save(plankton_sub, file = "data/00.plankton_sub.Rdata")
save(images_sub, file = "data/00.images_sub.Rdata")


## Proportion of each motility type ----
#--------------------------------------------------------------------------#
# Read all plankton data
plankton <- read_parquet("data/00.plankton_clean.parquet") 

# Read motility for each taxon
motility_types <- read_csv("data/raw/taxa_list.csv", show_col_types = FALSE) %>% 
  select(taxon = new_name, motility) %>% 
  drop_na() %>% 
  arrange(taxon)

# Compute proportion for each type of motility
motility_counts <- plankton %>% 
  count(taxon) %>% 
  filter(taxon != "Collodaria_colonial") %>% # ignore colonial Collodaria
  left_join(motility_types, by = join_by(taxon)) 

save(motility_counts, file = "data/00.motility_counts.Rdata")
