==> 06.intrataxa_distances_analysis.qmd <==
---
title: "Analyze intrataxonomic distances."
author: "Thelma Panaïotis"
format: 
  html:
    toc: true
    embed-resources: true
editor: visual
lightbox: true
execute:
  warning: false
  cache: true
  freeze: false
---

```{r set_up}
#| cache: false
source("utils.R")
```

Load computed distances.

```{r load}
# Intrataxa
load("data/05.intra_distances.Rdata")

# Overall
load("data/03.all_distances.Rdata")
```

Plot test stat

```{r}
df_intra %>% 
  ggplot() +
  geom_col(aes(x = taxon, y = test_stat, fill = n_obj)) +
  geom_hline(yintercept = df_all$test_stat, colour = "red") +
  scale_fill_viridis_c(trans = "log10") +
  theme(axis.text.x = element_text(angle = 45, hjust = 1))
```

```{r}
#| fig-column: body-outset
#| out-width: 100%
df_intra %>% 
  ggplot() + 
  geom_point(aes(x = n_img, y = test_stat, colour = taxon)) +
  scale_x_log10() + scale_y_log10()
```

Plot all distances.

```{r all_dist}
#| fig-column: screen-inset
#| out-width: 100%
#| fig-width: 12
#| fig-height: 5
df_intra_dist %>% 
  pivot_longer(dist:dist_rand) %>% 
  ggplot() +
  geom_density(aes(x = value, colour = name)) +
  facet_wrap(~taxon, scales = "free")
```

Plot all ECDF.

```{r all_ecdf}
#| fig-column: screen-inset
#| out-width: 100%
#| fig-width: 12
#| fig-height: 5
df_intra_dist %>% 
  pivot_longer(dist:dist_rand) %>% 
  ggplot() +
  stat_ecdf(aes(x = value, colour = name)) +
  facet_wrap(~taxon, scales = "free")
```
