==> 08.intertaxa_distances_analysis.qmd <==
---
title: "Analyze intertaxonomic distances."
author: "Thelma Panaïotis"
format: 
  html:
    toc: true
    embed-resources: true
editor: visual
lightbox: true
execute:
  warning: false
  cache: true
  freeze: false
---

```{r set_up}
#| cache: false
source("utils.R")
```

List saved distances and ream them.

```{r load}
## Inter distances
# list processed files
processed <- list.files("data", pattern = "07.", full.names = TRUE)

# load data
res <- sapply(processed, function(x) mget(load(x)), simplify = TRUE)

# get summary data (1st line)
df_inter <- res[1,] %>% bind_rows()
# get distances (2nd line)
df_inter_dist <- res[2,] %>% bind_rows()

# Ignore colonial Collodaria
#df_inter <- df_inter %>% filter(!str_detect(pair, "Collodaria_colonial"))
#df_inter_dist <- df_inter_dist %>% filter(!str_detect(pair, "Collodaria_colonial"))

## Overall distances
load("data/03.all_distances.Rdata")
```

Plot test stat

```{r}
df_inter %>% 
  separate(pair, into = c("t1", "t2"), sep = " - ") %>% 
  ggplot() + 
  geom_raster(aes(x = t1, y = t2, fill = test_stat)) +
  scale_fill_viridis_c() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1))
```

Plot number of objects.

```{r}
df_inter %>% 
  separate(pair, into = c("t1", "t2"), sep = " - ") %>% 
  ggplot() + 
  geom_raster(aes(x = t1, y = t2, fill = n_obj)) +
  scale_fill_viridis_c(trans = "log1p") +
  theme(axis.text.x = element_text(angle = 45, hjust = 1))
```

Test stat VS number of objects

```{r}
ggplot(df_inter) + 
  geom_point(aes(x = n_obj, y = test_stat, colour = str_detect(pair, "Collodaria_colonial"))) +
  labs(colour = "Collo_col") +
  scale_x_log10() + scale_y_log10()
```

Plot distributions

```{r}
df_inter_dist %>% 
  left_join(df_inter) %>% 
  filter(test_stat > 0.1) %>% 
  pivot_longer(dist:dist_rand) %>% 
  ggplot() + 
  geom_density(aes(x = value, colour = name)) +
  facet_wrap(~pair)
```

Plot ECDF

```{r}
df_inter_dist %>% 
  left_join(df_inter) %>% 
  filter(test_stat > 0.1) %>% 
  pivot_longer(dist:dist_rand) %>% 
  ggplot() + 
  stat_ecdf(aes(x = value, colour = name)) +
  facet_wrap(~pair)
```
