==> 10b.null_acantharea_analysis.qmd <==
---
title: "Analyze null Kuiper statistics for Acantharea."
author: "Thelma Panaïotis"
format: 
  html:
    toc: true
    embed-resources: true
editor: visual
lightbox: true
execute:
  warning: false
  cache: true
  freeze: false
---

```{r set_up}
#| cache: false
source("utils.R")
```

## Read saved distances

Read null distances and stats.

```{r read_null}
load("data/04b.null_ks.Rdata")
```

Read intra distances and stats.

```{r read_intra}
load("data/05b.intra_distances_ks.Rdata")

# Log transform n dist and kuiper stat
df_intra <- df_intra %>% 
  mutate(
    log_n_dist = log10(n_dist),
    log_kuiper_stat = log10(kuiper_stat)
  )

```

Read null Acantharea.

```{r read_acant}
load("data/10a.null_ks_n_dist_acant.Rdata")

# Round n_dist across sets for boxplot
null_ks_n_dist_acant <- null_ks_n_dist_acant %>% 
  mutate(log_n_dist = round(log_n_dist, digits = 1))

```

## Distribution of null Acantharea Kuiper statistics

```{r null_acant}
# Plot results, just Acant
ggplot() +
  geom_boxplot(data = null_ks_n_dist_acant, aes(x = log_n_dist, y = log_kuiper_stat)) +
  geom_point(data = df_intra %>% filter(taxon == "Acantharea"), aes(x = log_n_dist, y = log_kuiper_stat), colour = "red") +
  scale_x_continuous(labels = label_math(expr = 10^.x, format = force)) +
  scale_y_continuous(labels = label_math(expr = 10^.x, format = force)) +
  labs(x = "N distances", y = "Kuiper statistic")
```

## Compare with overall null Kuiper statistics

```{r plot_comp}

ggplot() + 
  geom_boxplot(data = null_ks_n_dist, aes(x = log_n_dist, y = log_kuiper_stat, group = n_dist), colour = "grey", width = 0.3) +
  geom_boxplot(data = null_ks_n_dist_acant, aes(x = log_n_dist, y = log_kuiper_stat), width = 0.3) +
  geom_point(data = df_intra %>% filter(taxon == "Acantharea"), aes(x = log_n_dist, y = log_kuiper_stat), colour = "red") +
  geom_polygon(data = null_ks_rib, aes(x = log_n_dist, y = y), alpha = 0.1) + 
  scale_x_continuous(labels = label_math(expr = 10^.x, format = force)) +
  scale_y_continuous(labels = label_math(expr = 10^.x, format = force)) +
  labs(x = "N distances", y = "Kuiper statistic") 
```

Black is null Acantharea, grey is the global null data, red is observed Acantharea distances.

::: {.callout-note appearance="simple" icon="false"}
Global null can be used not only for overall distances, but also for intra and inter subsets.
:::
